Update the cache when a parent is created
If multiple data bag items are specified on the command line to be
uploaded to a data bag that doesn't exist the first will be uploaded
but the next will fail with a 409 CONFLICT.  This is because each item
caches the fact that the parent doesn't exist and tries to create it.

This adds the first created parent to the cache such that subsequent
items do not try to recreate it.

Fix for issue: https://github.com/jkeiser/knife-essentials/issues/107